Dynamic Snek + added Settings.txt + float to int cast in board

+++ Engine/Board.cpp
@@ -7,8 +7,8 @@ Board::Board(Graphics& gfx, const GameSettings& settings)
 	: mGfx(gfx), mDimension(settings.GetTileSize()), mWidth(settings.GetBoardWidth()), mHeight(settings.GetBoardHeight())
 	, mContents(new ContentType[mWidth * mHeight]{ ContentType::Empty })
 	, mRandomX(0, mWidth - 1), mRandomY(0, mHeight - 1)
-	, mX(gfx.GetScreenRect().GetCenter().X - (mWidth * mDimension / 2))
-	, mY(gfx.GetScreenRect().GetCenter().Y - (mHeight * mDimension / 2))
+	, mX(static_cast<int>(gfx.GetScreenRect().GetCenter().X) - (mWidth * mDimension / 2))
+	, mY(static_cast<int>(gfx.GetScreenRect().GetCenter().Y) - (mHeight * mDimension / 2))
 {
 }
